requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
easydict==1.9
Flask==2.0.1
gast==0.3.3
gevent==21.1.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.37.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==2.0.1
Jinja2==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.3.0
pandas==1.1.5
Pillow @ file:///C:/ci/pillow_1617386319301/work
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pytesseract==0.3.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six @ file:///C:/ci/six_1605205426665/work
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==1.15.1
tensorflow-gpu==2.3.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020062301
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
